Timestamps: Major reorganization. Time zones implemented for date/time formats.
